Initial commit of MCUXpresso project files
This is the stock project for an MCUXpresso LPC822 C++ application.
This compiles and can be loaded onto the LPC822 MCU on the Teatotler
controller board. Further tweaks will follow to put this in a better
place under the sw/ directory. An ide/ directory is additionally
created to store these files.