Add PeriodicTask class
Add the PeriodicTask class to manage work done on a fixed time
interval. This class should provide the interface through which all
tasks in the system are run in the top-level control loop.

This commit removes ManagedDriver, since PeriodicTasks turned out to
be identical in its interface, and could apply equally to code in app/
and system/. It lives in util/ for this reason.